// reducers/SearchDialog/searchDialog.types.ts
export enum SearchDialogTypes {
  SetSearchText = "SET_SEARCH_TEXT",
  SetOpenDialog = "SET_OPEN_DIALOG",
  SetSearchResultList = "SET_SEARCH_RESULT_LIST",
  SetSearchByValue = "SET_SEARCH_VALUE",
  SetSortByValue = "SET_SORT_VALUE",
  SetSortByOrder = "SET_SORT_BY_ORDER",
  SetSortByEntity = "SET_SORT_BY_ENTITY",
  SetPageNumber = "SET_PAGE_NUMBER",
  SetRowsPerPage = "SET_ROWS_PER_PAGE",
}